COMPLETELY DONE WITH SCROLL AND SET-UP FOR BURRON POS!
everything is positions right and is very scalable to whatever amount if owners is needed. I will move forwards So it can be easily adapted with a DB and the rest of the pages! (The info that each Btn will call for is going to be printed in console btw)